chore: move Connector unit tests
